+++ src/SFA.DAS.Payments.EarningEvents.Messages.External/Earnings.cs
@@ -3,7 +3,7 @@ namespace SFA.DAS.Payments.EarningEvents.Messages.External
 {
     public class Earnings
     {
-        public int AcademicYear { get; set; }
+        public short AcademicYear { get; set; }
         public IEnumerable<PricePeriod> PricePeriods { get; set; }
     }
 }
